--- my-ability-uni/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id 'java'
-    id 'java-library'
-    id 'checkstyle'
-    id 'configure-jandex'
-}
-
-
-dependencies {
-    api enforcedPlatform(libs.quarkus.platform.bom)
-
-//    api project(':my-ability')
-    api project(':my-core')
-//    api project(':my-database-uni')
-
-}
