--- my-ability-uni/src/main/java/net/ximatai/muyun/ability/uni/ICURDUniAbility.java
@@ -1,4 +1,0 @@
-package net.ximatai.muyun.ability.uni;
-
-public interface ICURDUniAbility extends ICreateAbilityUni {
-}
