--- my-ability/build.gradle
@@ -1,12 +1,0 @@
-plugins {
-    id 'java'
-    id 'java-library'
-    id 'checkstyle'
-    id 'configure-jandex'
-}
-
-
-dependencies {
-    api enforcedPlatform(libs.quarkus.platform.bom)
-    api project(':my-core')
-}
