--- my-ability/src/main/java/net/ximatai/muyun/ability/ICURDAbility.java
@@ -1,4 +1,0 @@
-package net.ximatai.muyun.ability;
-
-public interface ICURDAbility extends ICreateAbility, IUpdateAbility, IDeleteAbility, ISelectAbility {
-}
