--- my-core/src/main/java/net/ximatai/muyun/ability/IDatabaseAbility.java
@@ -1,13 +1,0 @@
-package net.ximatai.muyun.ability;
-
-import net.ximatai.muyun.database.IDatabaseAccess;
-import net.ximatai.muyun.database.IDatabaseAccessStandard;
-
-public interface IDatabaseAbility {
-
-    IDatabaseAccess getDatabaseAccess();
-
-    default IDatabaseAccessStandard getDatabase() {
-        return (IDatabaseAccessStandard) getDatabaseAccess();
-    }
-}
