--- my-core/src/main/java/net/ximatai/muyun/ability/IDatabaseUniAbility.java
@@ -1,14 +1,0 @@
-package net.ximatai.muyun.ability;
-
-import net.ximatai.muyun.database.IDatabaseAccess;
-import net.ximatai.muyun.database.IDatabaseAccessUni;
-
-public interface IDatabaseUniAbility {
-
-    IDatabaseAccess getDatabaseAccess();
-
-    default IDatabaseAccessUni getDatabase() {
-        return (IDatabaseAccessUni) getDatabaseAccess();
-    }
-
-}
