+++ my-core/src/main/java/net/ximatai/muyun/ability/std/IDatabaseUniAbility.java
@@ -1,0 +1,14 @@
+package net.ximatai.muyun.ability.std;
+
+import net.ximatai.muyun.database.IDatabaseAccess;
+import net.ximatai.muyun.database.IDatabaseAccessUni;
+
+public interface IDatabaseUniAbility {
+
+    IDatabaseAccess getDatabaseAccess();
+
+    default IDatabaseAccessUni getDatabase() {
+        return (IDatabaseAccessUni) getDatabaseAccess();
+    }
+
+}
