+++ my-core/src/main/java/net/ximatai/muyun/ability/std/curd/std/ICURDAbility.java
@@ -1,0 +1,4 @@
+package net.ximatai.muyun.ability.std.curd.std;
+
+public interface ICURDAbility extends ICreateAbility, IUpdateAbility, IDeleteAbility, ISelectAbility {
+}
