+++ my-core/src/main/java/net/ximatai/muyun/ability/std/curd/uni/ICURDUniAbility.java
@@ -1,0 +1,4 @@
+package net.ximatai.muyun.ability.std.curd.uni;
+
+public interface ICURDUniAbility extends ICreateAbilityUni {
+}
