--- my-database-standard/build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'java'
-    id 'java-library'
-    id 'checkstyle'
-    id 'configure-jandex'
-}
-
-
-dependencies {
-    api enforcedPlatform(libs.quarkus.platform.bom)
-    implementation project(':my-database')
-
-    api "io.quarkus:quarkus-hibernate-orm"
-    api "io.quarkus:quarkus-jdbc-postgresql"
-}
