+++ my-database-std/build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'java'
+    id 'java-library'
+    id 'checkstyle'
+    id 'configure-jandex'
+}
+
+
+dependencies {
+    api enforcedPlatform(libs.quarkus.platform.bom)
+    implementation project(':my-database')
+
+    api "io.quarkus:quarkus-agroal"
+    api "io.quarkus:quarkus-hibernate-orm"
+    api "io.quarkus:quarkus-jdbc-postgresql"
+}
