+++ my-platform/build.gradle
@@ -8,7 +8,8 @@ plugins {
 dependencies {
     implementation project(':my-core')
 //    implementation project(':my-ability')
-//    implementation project(':my-database-standard')
-    implementation project(':my-ability-uni')
+//    implementation project(':my-database-std')
+//    implementation project(':my-ability')
+//    implementation project(':my-database-std')
     implementation project(':my-database-uni')
 }
