+++ settings.gradle
@@ -8,11 +8,11 @@ assert JavaVersion.current() >= JavaVersion.VERSION_21:
     "You must use at least Java 21 to build the project, you're currently using ${System.getProperty("java.version")}"
 
 include 'my-core'
-include 'my-ability'
-include 'my-ability-uni'
+//include 'my-ability'
+//include 'my-ability-uni'
 include 'my-platform'
 include 'my-database'
-include 'my-database-standard'
+include 'my-database-std'
 include 'my-database-uni'
 //include 'my-msg'
 include 'my-boot'
